Allocate humongous objects in young gen
This is a placeholder for an internal commit that is being upstreamed.  The
lines of code affected by the original internal commit were already changed
by some previously applied redundant commit operation.  I'm leaving this
empty commit in place to facilitate audits of the source history.

As originally drafted, this commit removed lines 124 and 125 of
heuristics/shenandoahHeuristics.cpp:
  // We never allocate humongous objects in young gen. (TODO: revisit this).
  assert(region->affiliation() != ShenandoahRegionAffiliation::YOUNG_GENERATION, "Humongous object found in young gen.");

and added the following after line 315 in shenandoahFreeSet.cpp:

    r->set_affiliation(req.affiliation());