Change definition of CollectionThresholdGarbagePercent
This is now defined to equal ShenandoahGarbageThreshold, which seems to have
a default value of 25.  The effect on running workloads is to choose more
regions for the collection set than was observed with the previous
configuration.

Also addressed several improvements in white space and comments.

The code now runs tier1 and hotspot_gc_shenandoah without regressions.  It
also succsessfully runs an Extremem stress test up until the point of
failure due to triggering of full GC (after 113 completed GC passes, including
two old-gen GC passes).