Clean up remembered set maintenance
Removed and replaced many direct accesses to ShenandoahCardTable.  Require code to use
the RememberedSet abstraction instead.  Removed unnecessary remembered set operations.
Tidied up other remembered set operations.  Inserted a bunch of instrumentation which will
need to be removed.  As currently implemented, this code works when I assume within
process clusters that every card is dirty.  But it crashes with assertion failures resulting
from failed forwarding of collection set objects when I only update references for dirty cards.
Next step is to try to figure out which dirty bit is being lost by finding an interesting
pointer on a card that is marked clean.  If we're lucky, the nature of this inconsistency
will direct us to the code that has failed to dirty the relevant cards.