Working with instrumentation
To make concurrent remembered set scanning work, I had to copy the remembered set
to a read-only copy at init-mark safepoint and then initialize the remembered set
to all clean.  The original plan was to simply swap pointers between a read-card-table
and a write-card-table.  I found that the existing mutator write-barrier implementations
were not able to adjust to a change in location for the write-card-table.

Before upstreaming, I need to remove instrumentation and then confirm that I have not
broken anything.