Assure that memory holding weak references from old to young is marked dirty
During remembered set scanning, "weak references" are not passed to mark_through_ref,
so dirty bits in the read-card-table may not be copied to the write-card-table.
In the case that a Reference object is dropped, meaning that its weak reference is
not going to be nulled, we dirty the associated card within the write-card-table.
Patch provided by William Kemper.