Progress in debugging concurrent old GC
This commit includes code to enable mark-through-ref during mixed evacuations,
to use mark bits to restrict scanning of remembered sets during mixed evacuations.
This commit also includes a bunch of instrumentation that will eventually be removed.
But I'm leaving it in place for the moment because there are still some bugs that
I'm trying to chase down.