Promote full regions during concurrent evac
Prior to this patch, full region promotion was handled as part of the final-update-refs safepoint, resulting in occasional
pauses in excess of 500 ms.  This change disables atomic region promotion under safepoint.  Instead, we encourage regions
that have reached the tenure age to be included in the collection set.  Promotion of humongous regions, which does not copy
data but does have to create remembered set information, is also performed during the concurrent evacuation phase.