Progress towards eliminating degenerated and full GC with large workload
As committed, this code is crashing when OLD is triggered.  I'm
committing these changes in order to rebuild on an rr-enabled host where
I can debug further.

This includes the following changes, which will most likely be
upstreamed as a sequence of independent pull requests:

1. We were resetting or incrementing the age of regions in final-mark.
This was the wrong place to do this, because it ignored the allocation
of objects during evacuation, including the allocations for survivor
objects within survivor regions.  I've moved this code to final update
references.

2. I've changed the mechanism for biasing young-gen regions that have
reached promotion age and changed the default values.  I find that I
need to multiply the garbage by 16 or more to get reasonable behavior,
whereas the original implementation was multiplying the garbage by only
1.5.

3. When choosing collection sets,  we were including regions that had
very little garbage (e.g. only 16 bytes of garbage!) if we had not yet
reached our "minimum amount of collected garbage" heuristic threshold.
I've removed this heuristic as it is wasteful and counterproductive.
I've left a comment that we might want to revisit this design in the future.

4. When choosing collection sets, in the case that we had met our our
"minimum amount of collected garbage" heuristic, we were excluding regions
that had reached promotion age if the amount of garbage was less than
ShenandoahGarbageThreshold (default 25%).  I'm in the process of changing
this to include all promotion-age regions that fit within the ShenandoahEvacReserve
(default 5% of generation capacity), still prioritizing so regions with more
garbage are inserted into the collection set before regions with less garbage.
I had observed that the collection set was often ignoring large numbers of
promotion-age regions even while it was well below its potential full evacuation
reserve (I saw cases where the collection set was less than 50% of its evacuation
capacity while over 200 promotion-age regions were left out of the collection set).

5. I also inserted some "improvements" into the should_start_gc() heuristics.  If
live-memory in generation is increasing or if previous gc was degenerated, I "scale"
my prediction of expected gc execution time so that we trigger the next gc sooner
than with original heuristic implementation.

6.  Fixed up full GC so that it properly ages and promotes objects.  While the fixes
above are designed to greatly reduce the dependency on full GC, we need full GC to
behave better if it ever is invoked.

One area known to require additional attention is the implementation of
memory budget enforcment for old-gen evacuation and promotion.  Some of
the code in place assumes we have old-gen GCLABs.  That was a
misunderstanding on my part.  All old-gen allocations, both for
evacuation and for promotion, use PLAB buffers.  I've started work on
fixing this, as represented by a few declarations inserted into
ShenandoahThreadLocal class.