Simplify crossing map implementation
The crossing map maintains a record of the first and last objects
starting within each remembered set card's range of memory addresses.
The previous implementations used bit-shifts.  It was detected during
testing that the c++ compiler was emitting incorrect code in certain
in-line expansions of the register_object() method.  gdb reports that
there are 56 distinct contexts holding the expansion of this function.
This simplified implementation does not exhibit the same broken
behavior.