Fix invocations of get_next_marked_addr
There were a few places in existing code that misused this service.
get_next_marked_addr expects a start and end arguments to delineate
the range of addresses within which to search for the next marked
object.  If the next marked object is found at the start address, it
returns the start address.  Some existing code assumed it would not find
an object at this address, but would find the "next" object after this
one.  This patch corrects the misunderstanding and fixes other erroneous
usage based on the mistaken belief that get_next_marked_addr returns
nullptr if there are no marked objects within the range.  Instead, it
returns the value of the end argument.