Reserve certain old-gen regions to support compaction of old-gen
These regions can be used only for compaction of old-gen memory.
Following compaction of old-gen memory, at least the same number of
regions will become free as a result of compaction within old-gen
memory.  Promotions will be rejected (and deferred until a subsequent
evacuation phase) if they would need to occupy regions that are reserved
for compaction of old-gen memory.