Make promotion budget more generous
The reason to enforce a promotion budget is to make more memory
available for loaning to young-gen allocations during evacuation and
update-refs phases of gc.  A second reason to enforce a promotion budget
is to assure that old-gen memory does not get so full that there is no
memory available into which to copy objects during old-gen evacuations.

With this patch, if it is available, we set the promotion budget large
enough to hold all of the live objects contained within the young-gen
collection set.

Manually resolved merge conflicts and brought in some additional context
from the development branch.