Performance improvements for object promotion
Reset promoted	expended at start of concurrent	GC rather than end.  This corrects a bug that manifests when concurrent
GC degenerates and we don't finish in the normal way.  Following a degenerated GC, we were failing to reset promoted
expended.

Fix up the remediation code that executes when PLAB allocation fails.  We had previously failed	to provide any special
handling for threads that do not have PLABs so objects encountered by such a thread were always evacuated to young so
they could be promoted in a subsequent GC pass.	 While it was relatively harmless to leave these rare objects in young-gen
memory, a more	significant impact was that this event would be logged as promotion failure and this would trigger immediate
execution of an	old-gen	collection, even when old-gen memory is	plentiful.  Now, if a thread that does not have a PLAB
needs to promote an object, we use a shared allocation to obtain memory for the object's copy.  There seems to be a single
JVM thread that is set up without a PLAB and this thread occassionally stumbles onto objects that need to be promoted.

Be more	generous in the	computation of promotion reserve.  The miserly budget previously set aside was causing
promotions to fail, resulting in significant increase in attempts at shared-memory allocations and causing older
objects	to accumulate in young-gen.  During subsequent GC passes, these	objects	would once again be likely to fail
to promote, causing a a	repeat of the problems just described.

Insert code to account for round-off errors that occur in calculations and enforcement of promotion budgets.  The
previous failure to account for	these round-off	errors was resulting in	wraparound arithmetic on the size_t type,
causing	promotion budget to be extremely large in some rare cases.  This would result in depletion of old-gen space
to the extent that subsequent evacuations of old-gen could not be performed.

Add logging reports when promotions fail.

Allow memory set aside to fulfill the minimum evacuation reserve in old-gen to be loaned to young-gen if the memory
is not needed in the current pass to hold old-gen evacuations.