Use smaller ChunkSize for remembered set scanning
Prior to this change, the first half of remembered set scanning assigned
entire regions at a time to individual GC worker threads.  This was
observed to result in too much imbalance between the work performed by
individual threads.  With this fix, the maximum ChunkSize is the smaller
of the region size and 1mb.  This fix was observed to improve worst-case
remembered set scanning by 50% on a workload that had approximately 10GB
old-gen and 52 GB heap size.