8298647: GenShen require heap size 2MB granularity
Generational Shenandoah requires 2MB granularity in order for card
tables to cover the allocated heap. Each byte in a page of card table
represents 512 heap bytes. As card tables are allocated 4KB at a time,
4KB * 512 = 2MB.

There is a circular dependency between the region calculations and the
heap size calculations. This unconditionally rounds up the heap size to
2MB. It might be preferable to do this only when generational mode is
enabled.

Running with:
  java -Xlog:gc*=trace -XX:+UseShenandoahGC -mx495m \
       -XX:ShenandoahGCMode=generational -version

on a debug build is sufficient to reproduce this problem.