Increase size of largest ShenandoahRegionChunk
Size was 1 MB.  New size is 4 MB.  This decreases the number of chunks
required to span all of memory, which decreases the number of
synchronizations required to obtain chunk assignments. which decreases
the number of synchronizations required to obtain chunk assignments.
which decreases the number of synchronizations required to obtain chunk
assignments. which decreases the number of synchronizations required to
obtain chunk assignments.