Enforce bounds on regions per generation
This adds code to ShenandoahVerifier, adds specific log messages, and
introduces new code to force compliance with bounds on the number of
heap regions affiliated with each generation.  This was accidentally not
enforced previously.  The observed consequence of failing to enforce
this constraint was that allocations were failing even though the total
bytes available within particular regions was very large (over 10 GBytes
in some cases).