Do not enable_plab_promotions in try_evacuate_object
The trigger for this was too liberal, and we ended up thrashing between
disabling and enabling plab promotions.  This performs much better with
this change.  However, once a thread's plab promotions have been
disabled, that thread will never again be able to perform plab
promotions.

I need to add code that re-enabled plab promotions at the start of the
evacuation cycle.