Account for regular regions promoted in place promptly
In the previous implementation, accounting was deferred until the end
the evacuation and updating cycles.  However, the affiliation of the
region was changed at the time the region was promoted.  The problem
that arose was that PLABs and shared allocations can be taken from the
newly promoted regions, resulting in usage that exceeds the
old-generationi capacity and exceeds the span of the regions that are
affiliated with the old generation.  Then, availability in old becomes
"negative" because available is calculated as capacity - used.  But
negative is really ginormous, because the type of available is unsigned
size_t.  This patch fixes these problems.