When (generational) GC discovers j.l.Reference:s in both generations, then subsequent processing of the discovered list may create new old to young pointers which the young collector needs to know about. This fix takes care of marking cards for that case.
The issue was revealed during a verification failure and subsequently
verified by means of a check in the reference processing code to detect
the condition when the new cross generational pointer was being created
without the associated card-marking. This typically happens when
discovery is in both generations -- for young collections such Reference
objects in the old generation are discovered during the card scanning,
and for global collections when objects are marked. The situation is
typically rare in the case of young collections because the Reference
objects typically lie on dirty cards and hold pointers to referents in
the young generation.