Remove infrastructure to adjust generation capacities for loans from old
A previous mechanism would loan memory from old to young during the
evacuation and update-refs phases of GC.  This approach had limited
value because the loans had to be repaid immediately upon completion of
the update-refs phase.  A newer approach dynamically adjusts the size of
old generation, consistently keeping it as small as possible.  This
allows the maximum amount of allocation to be performed between
consecutive young-gc efforts.