Overwrite update_water_mark in ShenandoahHeapRegion::promote_in_place
While a region waits to be promoted in place, its value of top equals
end.  This is a temporary arrangement, to prevent young-gen objects
from being allocated within this region before it is promoted.  Once
the region is promoted, its value of top() is adjusted to its previous
value.  The region's update_watermark value was likely set when the
region had its artificially high value of top.  So we fix it here.