Changes to simplify MU accounting
This represents a small part of the expand-old-on-demand branch.  By
itself, this may result in some performance regression because we no
longer use MU (mutator utilization) reports to guide generation sizing
decisions.  On the other hand, the guidance provided by MU metrics does
not seem to be a very accurate predictor of ideal generation sizes.