Improve handoff of generation sizing impacts
At the end of a completed GC cycle, we make adjustments in the
generation sizes.  These adjusted are based on the number of trash
regions (the size of the just-evacuated collection set) and the number
of regions needed within old-gen to support any mixed evacuation effort
or promotion that is anticipated to happen during the next GC cycle.
Recycling of trashed regions and transfer of region membership between
generations happens after we rebuild the free set.  This commit assures that
the free set rebuild process know how much memory will eventually be
available in each generation after we finish reclaiming trashed regions
and transfering regions between generations.