Experiment with alternative defaults
If we decide to change these default values, we will need to arrange for
GenShen default to be different than SingleShen default.  Setting these
defaults to experiment with our performance pipeline.

Note that we need to sample more frequently if we are going to rely on
ability to quickly respond to changes in allocation rate.  When we have
the ability to quickly respond to changes in allocation rate, it is more
important to forget history more quickly.  Otherwise, we have overly
conservative triggering because we do not recognize that occassionally
long cycles are not representative of typical performance.