Do not align_up size of object allocated in plab slow
Align the plab size, but not the size of the object to be allocated within the plab.
If we increase the size of the allocated object, this creates confusion in that the
object's size reported by its header does not span all of the bytes allocated for the
object, so the memory is not parsable during remembered set scanning.